+++ src/components/TodoList/TodoList.tsx
@@ -1,9 +1,9 @@
-import cn from 'classnames';
 import { useContext } from 'react';
 import { CSSTransition, TransitionGroup } from 'react-transition-group';
 import { Todo } from '../../types/Todo';
-import { TodoItem } from '../TodoField';
+import { TodoField } from '../TodoField';
 import { StateContext } from '../../State/State';
+import { TodoItem } from '../TodoItem';
 
 type Props = {
   todos: Todo[];
@@ -21,7 +21,7 @@ export const TodoList: React.FC<Props> = ({ todos }) => {
             timeout={300}
             classNames="item"
           >
-            <TodoItem todo={todo} key={todo.id} />
+            <TodoField todo={todo} key={todo.id} />
           </CSSTransition>
         ))}
 
@@ -31,38 +31,7 @@ export const TodoList: React.FC<Props> = ({ todos }) => {
             timeout={300}
             classNames="temp-item"
           >
-            <div data-cy="Todo" className="todo is-active">
-              <label className="todo__status-label">
-                <input
-                  data-cy="TodoStatus"
-                  type="checkbox"
-                  className="todo__status"
-                />
-              </label>
-
-              <span data-cy="TodoTitle" className="todo__title">
-                {tempTodo.title}
-              </span>
-
-              <button
-                type="button"
-                className="todo__remove"
-                data-cy="TodoDelete"
-              >
-                ×
-              </button>
-
-              {/* 'is-active' class puts this modal on top of the todo */}
-              <div
-                data-cy="TodoLoader"
-                className={cn('modal overlay', {
-                  'is-active': true,
-                })}
-              >
-                <div className="modal-background has-background-white-ter" />
-                <div className="loader" />
-              </div>
-            </div>
+            <TodoItem todo={tempTodo} />
           </CSSTransition>
         )}
       </TransitionGroup>
